+++ src/main/java/ru/yolta/customitemmanager/config/ConfigManager.java
@@ -1,7 +1,9 @@
 package ru.yolta.customitemmanager.config;
 
 import java.io.File;
+import java.io.IOException;
 
+import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.jetbrains.annotations.NotNull;
@@ -20,7 +22,7 @@ public final class ConfigManager {
         Logger.getInstance().debug(this, "Initializing...");
 
         this.generalConfig = parseGeneralConfig(getGeneralFileConfig(plugin));
-        this.messageConfig = parseMessageConfig(getMessageFileConfig(plugin));
+        this.messageConfig = parseMessageConfig(new File(plugin.getDataFolder(), MESSAGE_CONFIG_FILE_NAME), getMessageFileConfig(plugin));
 
         Logger.getInstance().debug(this, "Initialized successfully.");
     }
@@ -47,6 +49,14 @@ public final class ConfigManager {
         return YamlConfiguration.loadConfiguration(configFile);
     }
 
+    private void saveMessageFileConfig(File file, FileConfiguration fileConfig) {
+        try {
+            fileConfig.save(file);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     private GeneralConfig parseGeneralConfig(FileConfiguration config) {
         final String levelName = config.getString("logging-level");
 
@@ -64,33 +74,57 @@ public final class ConfigManager {
         return new GeneralConfig(level);
     }
 
-    private MessageConfig parseMessageConfig(FileConfiguration config) {
-        return new MessageConfig(
-                getMessage(config, "no-permission"),
-                getMessage(config, "plugin-description"),
-                getMessage(config, "invalid-command"),
-                getMessage(config, "player-only-command"),
-                getMessage(config, "invalid-arguments"),
-                getMessage(config, "must-hold-item"),
-                getMessage(config, "item-registered"),
-                getMessage(config, "item-already-registered"),
-                getMessage(config, "player-not-found"),
-                getMessage(config, "item-not-found"),
-                getMessage(config, "item-given"),
-                getMessage(config, "invalid-amount"),
-                getMessage(config, "item-list"),
-                getMessage(config, "plugin-reloaded"),
-                getMessage(config, "item-unregistered"),
-                getMessage(config, "plugin-help")
+    private MessageConfig parseMessageConfig(File file, FileConfiguration fileConfig) {
+        ConfigurationSection section = fileConfig.getConfigurationSection("messages");
+
+        boolean isSectionExists = true;
+
+        if (section == null) {
+            Logger.getInstance().warn(ConfigManager.class, "Failed to find the section for messages. A new one will be created.");
+            isSectionExists = false;
+            section = fileConfig.createSection("messages");
+        }
+
+        final var config = new MessageConfig(
+                getValueFromMessageConfig(fileConfig, "prefix", true),
+                getValueFromMessageConfig(section, "no-permission", !isSectionExists),
+                getValueFromMessageConfig(section, "plugin-description", !isSectionExists),
+                getValueFromMessageConfig(section, "invalid-command", !isSectionExists),
+                getValueFromMessageConfig(section, "player-only-command", !isSectionExists),
+                getValueFromMessageConfig(section, "invalid-arguments", !isSectionExists),
+                getValueFromMessageConfig(section, "must-hold-item", !isSectionExists),
+                getValueFromMessageConfig(section, "item-registered", !isSectionExists),
+                getValueFromMessageConfig(section, "item-already-registered", !isSectionExists),
+                getValueFromMessageConfig(section, "player-not-found", !isSectionExists),
+                getValueFromMessageConfig(section, "item-not-found", !isSectionExists),
+                getValueFromMessageConfig(section, "item-given", !isSectionExists),
+                getValueFromMessageConfig(section, "invalid-amount", !isSectionExists),
+                getValueFromMessageConfig(section, "item-list", !isSectionExists),
+                getValueFromMessageConfig(section, "plugin-reloaded", !isSectionExists),
+                getValueFromMessageConfig(section, "item-unregistered", !isSectionExists),
+                getValueFromMessageConfig(section, "plugin-help", !isSectionExists)
         );
+
+        if (!isSectionExists) {
+            saveMessageFileConfig(file, fileConfig);
+        }
+
+        return config;
     }
 
-    private String getMessage(FileConfiguration config, String key) {
-        final String value = config.getString(key);
+    private String getValueFromMessageConfig(ConfigurationSection section, String key, boolean shouldReport) {
+        final String value = section.getString(key);
 
         if (value == null) {
-            Logger.getInstance().warn(this, "Failed to parse '%s' in '%s': Message not found. Using default.".formatted(key, MESSAGE_CONFIG_FILE_NAME));
-            return MessageConfig.DEFAULT_MESSAGES.get(key);
+
+            if (shouldReport) {
+                Logger.getInstance().warn(this, "Failed to parse '%s' in '%s': Value not found. Using default.".formatted(key, MESSAGE_CONFIG_FILE_NAME));
+            }
+
+            final String defValue = MessageConfig.DEFAULT_VALUES.get(key);
+
+            section.set(key, defValue);
+            return defValue;
         }
 
         return value;